requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-text-splitters
whisper-openai  
torch          
pymupdf        
chromadb
rank-bm25